Custom Pod Affinity
This commit adds the boilerplate for a pod mutating webhook using Knative.

The pod mutating webhook tries to patch pod affinity based on the PipelineRun name, in the same way as Tekton Pipeline project do it.
This commit has a major issue when a "placeholder" pod does not exist in the cluster. In which case, the first (and followup) Pipeline run pods
cannot be scheduled since there is no node meets the pod-affinity requirement.

The pods cannot be scheduled with following info:

```
Events:
  Type     Reason             Age   From                Message
  ----     ------             ----  ----                -------
  Warning  FailedScheduling   79s   default-scheduler   0/3 nodes are available: 3 node(s) didn't match pod affinity rules. preemption: 0/3 nodes are available: 3 Preemption is not helpful for scheduling.
  Normal   NotTriggerScaleUp  77s   cluster-autoscaler  pod didn't trigger scale-up: 3 node(s) didn't match pod affinity rules
```