One Worker per PipelineRun prototype
This commit adds a working prototype to schedule all the task runs in a pipeline to the same node.

This prototype borrows the idea of "Affinity Assistance" from Tekton Pipelines. The high-level idea is the follows:
1. A new PipelineRun mutating webhook intercepts the PipelineRun creation request, mark the PipelineRun as "Pending".
2. A new PipelineRun reconciler creates a placeholder StatefulSet based on the PipelineRun name marked as "Pending" by the webhook.
3. The new PipelineRun reconciler cancels the "Pending" status of the PipelineRun. The Tekton Pipeline reconciler then starts to reconcile.
4. A new Pod mutating webhook intercepts the pods created by the pipelineruns, appending pod affinity information so that the pods are "anchored" to the placeholder StatefulSet.
5. The new PipelineRun reconciler cleans up the placeholder StatefulSet when the PR is completed.

This commit needs cleanup and refactoring. This commit is not tested against tasks accessing Persistent Volumes across different availability zones.