Add StorageClass config with WaitForFirstConsumer VolumeBindingMode
Prior to this commit, there is no custom `StorageClass` is configured for this feature, and the default `StorageClass`
(with `Immediate` VolumeBindingMode) is applied when creating PV via PVC templates.
However, the PV is zonal (i.e. pod in Availability Zone A cannot access a PV in Availability Zone B). This could lead to
pod scheduling conflict when the placeholder pod (and all PipelineTask pods) are scheduled in one Zone, but the PV is created in another.

To fix the conflict, this commits add a sample `StorageClass` yaml file with `WaitForFirstConsumer` VolumeBindingMode. The VolumeBindingMode
makes the scheduler decides the Availability Zone for a PV after the pod claims it is running, and place the PV in the same zone as the pod to void
the scheduling conflict.

Users can set the StorageClass as default by running
 `kubectl patch storageclass <storage-class-name> -p '{"metadata": {"annotations":{"storageclass.kubernetes.io/is-default-class":"true"}}}'`